appletb: Fix missing key-release events when fn key changed.
We were not properly tracking whether the key had been translated when
it was pressed, instead always looking at the current state of the fn
key. This would result in an untranslated key sent on release if the
fn key was released before releasing the touchbar key, and visa versa,
which would then be (rightfully) ignored by the key processing, thereby
resulting in lost key releases.

This fixes #5.