applespi: Ignore GPE's from wakeup.
If system wakeup via keyboard/touchpad is enabled, then the GPE is
(re)enabled by the core just before going to sleep, and consequently
when a key is pressed to wake up the system the key-release generates
another GPE which our handler then receives (before the core can
re-disable the GPE). Since the SPI device is not ready, this results
in an ugly error message in dmesg.

So we now remember when we're suspended and just ignore all GPE's while
in that state.